+++ R/snippet.R
@@ -56,9 +56,11 @@
 #' a value must be provided explicitly via \code{paper} when rendering to png.
 #' Otherwise for png outputs the height is cropped automatically rather than
 #' remaining a full page. See \code{lilypond} for details.
+#'
 #' Passing arguments to \code{header} can completely or partially prevent
-#' cropping, which must then be done manually with \code{linewidth}. This is
-#' all based on underlying LilyPond behavior.
+#' cropping in both direction, which must then be done manually with
+#' \code{linewidth} and \code{textheight}. This is all based on underlying
+#' LilyPond behavior.
 #'
 #' @param music a music object.
 #' @param file character, output file ending in .pdf or .png for sheet music or
@@ -79,6 +81,9 @@
 #' @param colors a named list of LilyPond element color global overrides. See
 #' \code{lilypond} for details.
 #' @param transparent logical, transparent background, png only.
+#' @param keep_ly logical, keep the intermediary LilyPond file.
+#' @param simplify logical, uses \code{simplify_phrase} to convert to simpler,
+#' more efficient LilyPond syntax.
 #'
 #' @return nothing returned; a file is written.
 #' @export
@@ -111,54 +116,57 @@
 render_music <- function(music, file, staff = "treble", tuning = "standard",
                          no_tab = TRUE, string_names = NULL, header = NULL,
                          paper = NULL, midi = FALSE, colors = NULL,
-                         transparent = FALSE){
+                         transparent = FALSE, keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = staff, no_tab = no_tab) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, colors, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_tc <- function(music, file, header = NULL, paper = NULL,
-                            midi = FALSE, colors = NULL, transparent = FALSE){
+                            midi = FALSE, colors = NULL, transparent = FALSE,
+                            keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(music_staff = "treble", no_tab = TRUE) %>%
     score() %>%
-    tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, FALSE,
-        colors, TRUE, transparent, FALSE)
+    tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, colors,
+        TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_bc <- function(music, file, header = NULL, paper = NULL,
-                            midi = FALSE, colors = NULL, transparent = FALSE){
+                            midi = FALSE, colors = NULL, transparent = FALSE,
+                            keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(music_staff = "bass", no_tab = TRUE) %>%
     score() %>%
-    tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, FALSE,
-        colors, TRUE, transparent, FALSE)
+    tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, colors,
+        TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_tab <- function(music, file, staff = NA, tuning = "standard",
                              string_names = NULL, header = NULL, paper = NULL,
-                             midi = FALSE, colors = NULL, transparent = FALSE){
+                             midi = FALSE, colors = NULL, transparent = FALSE,
+                             keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = staff) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, colors, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 #' @export
@@ -166,28 +174,30 @@ render_music_tab <- function(music, file, staff = NA, tuning = "standard",
 render_music_guitar <- function(music, file, tuning = "standard",
                                 string_names = NULL, header = NULL,
                                 paper = NULL, midi = FALSE,
-                                colors = NULL, transparent = FALSE){
+                                colors = NULL, transparent = FALSE,
+                                keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = "treble_8") %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, colors, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_bass <- function(music, file, tuning = "bass",
                               string_names = NULL, header = NULL, paper = NULL,
-                              midi = FALSE, colors = NULL, transparent = FALSE){
+                              midi = FALSE, colors = NULL, transparent = FALSE,
+                              keep_ly = FALSE, simplify = TRUE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = "bass_8") %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, colors, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, keep_ly, simplify, FALSE)
 }
 
 .paper_snippet <- function(x){