+++ NEWS.md
@@ -31,6 +31,7 @@
 * No more need (or support for) the `s`-suffix string numbering. All instances of single string number inputs are assumed starting string and any additional strings are inferred consecutively.
 * Improvements to `plot_fretboard` (renamed from `fretboard_plot`) and added wrapper function `plot_chord` for more convenient chord diagrams.
 * Minor bug fixes.
+* Significantly simplified LilyPond syntax for generated LilyPond files by adding `simplify_phrase` and the new (default) argument to `lilypond`, `simplify = TRUE`, which is also used by associated `render_*` functions.
 * Thank you to [fnord-repeater](https://github.com/fnord-repeater) for several helpful suggestions and insights as well as example code that helped to make the transcription pipeline better.
 * Thank you to [Han Oostdijk](https://github.com/HanOostdijk) for additional bug fixes and improvements to the code for the transcription pipeline.